website: base: Update icon size and format for modern devices.
This change makes the website icon 192x192 pixels, so that it looks
better in mobile devices; for example, when a user creates a launcher
icon to put in the home screen.

An SVG version of the icon is provided for devices that support it.

The bigger size icon is usually scaled down in Web clients that
display smaller icons (16x16 or 32x32 px).

* website/apps/base/templates/theme.scm (theme): Use the new PNG and
  SVG icons.

* website/apps/blog/templates/feed.scm (atom-feed-t): Use new PNG
  icon.

* website/static/base/img/favicon.png: Delete it.

* website/static/base/img/icon.png: Update background color and size.

* website/static/base/img/icon.svg: Add it.